chore(node): reduce usage of `Function`
