+++ types/node/assert.d.ts
@@ -33,17 +33,17 @@ declare module "assert" {
         function deepStrictEqual(actual: any, expected: any, message?: string | Error): void;
         function notDeepStrictEqual(actual: any, expected: any, message?: string | Error): void;
 
-        function throws(block: Function, message?: string | Error): void;
-        function throws(block: Function, error: RegExp | Function | Object | Error, message?: string | Error): void;
-        function doesNotThrow(block: Function, message?: string | Error): void;
-        function doesNotThrow(block: Function, error: RegExp | Function, message?: string | Error): void;
+        function throws(block: () => any, message?: string | Error): void;
+        function throws(block: () => any, error: RegExp | Function | Object | Error, message?: string | Error): void;
+        function doesNotThrow(block: () => any, message?: string | Error): void;
+        function doesNotThrow(block: () => any, error: RegExp | Function, message?: string | Error): void;
 
         function ifError(value: any): void;
 
-        function rejects(block: Function | Promise<any>, message?: string | Error): Promise<void>;
-        function rejects(block: Function | Promise<any>, error: RegExp | Function | Object | Error, message?: string | Error): Promise<void>;
-        function doesNotReject(block: Function | Promise<any>, message?: string | Error): Promise<void>;
-        function doesNotReject(block: Function | Promise<any>, error: RegExp | Function, message?: string | Error): Promise<void>;
+        function rejects(block: (() => Promise<any>) | Promise<any>, message?: string | Error): Promise<void>;
+        function rejects(block: (() => Promise<any>) | Promise<any>, error: RegExp | Function | Object | Error, message?: string | Error): Promise<void>;
+        function doesNotReject(block: (() => Promise<any>) | Promise<any>, message?: string | Error): Promise<void>;
+        function doesNotReject(block: (() => Promise<any>) | Promise<any>, error: RegExp | Function, message?: string | Error): Promise<void>;
 
         const strict: typeof internal;
     }
