+++ types/node/http.d.ts
@@ -1,8 +1,8 @@
 declare module "http" {
     import * as events from "events";
-    import * as net from "net";
     import * as stream from "stream";
     import { URL } from "url";
+    import { Socket, Server as NetServer } from "net";
 
     // incoming headers will never contain number
     interface IncomingHttpHeaders {
@@ -85,7 +85,7 @@ declare module "http" {
         timeout?: number;
         setHost?: boolean;
         // https://github.com/nodejs/node/blob/master/lib/_http_client.js#L278
-        createConnection?: (options: ClientRequestArgs, oncreate: (err: Error, socket: net.Socket) => void) => net.Socket;
+        createConnection?: (options: ClientRequestArgs, oncreate: (err: Error, socket: Socket) => void) => Socket;
     }
 
     interface ServerOptions {
@@ -95,7 +95,7 @@ declare module "http" {
 
     type RequestListener = (req: IncomingMessage, res: ServerResponse) => void;
 
-    class Server extends net.Server {
+    class Server extends NetServer {
         constructor(requestListener?: RequestListener);
         constructor(options: ServerOptions, requestListener?: RequestListener);
 
@@ -120,7 +120,7 @@ declare module "http" {
         sendDate: boolean;
         finished: boolean;
         headersSent: boolean;
-        connection: net.Socket;
+        connection: Socket;
 
         constructor();
 
@@ -142,8 +142,8 @@ declare module "http" {
 
         constructor(req: IncomingMessage);
 
-        assignSocket(socket: net.Socket): void;
-        detachSocket(socket: net.Socket): void;
+        assignSocket(socket: Socket): void;
+        detachSocket(socket: Socket): void;
         // https://github.com/nodejs/node/blob/master/test/parallel/test-http-write-callbacks.js#L53
         // no args in writeContinue callback
         writeContinue(callback?: () => void): void;
@@ -153,27 +153,27 @@ declare module "http" {
 
     // https://github.com/nodejs/node/blob/master/lib/_http_client.js#L77
     class ClientRequest extends OutgoingMessage {
-        connection: net.Socket;
-        socket: net.Socket;
+        connection: Socket;
+        socket: Socket;
         aborted: number;
 
         constructor(url: string | URL | ClientRequestArgs, cb?: (res: IncomingMessage) => void);
 
         readonly path: string;
         abort(): void;
-        onSocket(socket: net.Socket): void;
+        onSocket(socket: Socket): void;
         setTimeout(timeout: number, callback?: () => void): this;
         setNoDelay(noDelay?: boolean): void;
         setSocketKeepAlive(enable?: boolean, initialDelay?: number): void;
     }
 
     class IncomingMessage extends stream.Readable {
-        constructor(socket: net.Socket);
+        constructor(socket: Socket);
 
         httpVersion: string;
         httpVersionMajor: number;
         httpVersionMinor: number;
-        connection: net.Socket;
+        connection: Socket;
         headers: IncomingHttpHeaders;
         rawHeaders: string[];
         trailers: { [key: string]: string | undefined };
@@ -195,7 +195,7 @@ declare module "http" {
          * Only valid for response obtained from http.ClientRequest.
          */
         statusMessage?: string;
-        socket: net.Socket;
+        socket: Socket;
         destroy(error?: Error): void;
     }
 
@@ -226,8 +226,12 @@ declare module "http" {
     class Agent {
         maxFreeSockets: number;
         maxSockets: number;
-        sockets: any;
-        requests: any;
+        readonly sockets: {
+            readonly [key: string]: Socket[];
+        };
+        readonly requests: {
+            readonly [key: string]: IncomingMessage[];
+        };
 
         constructor(opts?: AgentOptions);
 
@@ -249,7 +253,6 @@ declare module "http" {
 
     function createServer(requestListener?: RequestListener): Server;
     function createServer(options: ServerOptions, requestListener?: RequestListener): Server;
-    function createClient(port?: number, host?: string): any;
 
     // although RequestOptions are passed as ClientRequestArgs to ClientRequest directly,
     // create interface RequestOptions would make the naming more clear to developers
