+++ types/node/test/cluster.ts
@@ -1,0 +1,11 @@
+import * as cluster from 'cluster';
+
+cluster.fork();
+Object.keys(cluster.workers).forEach(key => {
+    const worker = cluster.workers[key];
+    if (worker && worker.isDead()) {
+        console.log('worker %d is dead', worker.process.pid);
+    }
+});
+
+cluster.on('setup', (settings: cluster.ClusterSettings) => {});
