+++ types/node/test/http2.ts
@@ -326,8 +326,6 @@ import { URL } from 'url';
         response.end('', 'utf8', () => {});
         response.end(Buffer.from([]));
         response.end(Buffer.from([]), () => {});
-        response.end(Buffer.from([]), 'utf8');
-        response.end(Buffer.from([]), 'utf8', () => {});
 
         request.on('aborted', (hadError: boolean, code: number) => {});
         request.on('close', () => {});
