+++ types/node/tsconfig.json
@@ -42,25 +42,30 @@
         "vm.d.ts",
         "worker_threads.d.ts",
         "zlib.d.ts",
-        "node-tests.ts",
 
+        "node-tests.ts",
+        "test/assert.ts",
         "test/buffer.ts",
         "test/child_process.ts",
+        "test/cluster.ts",
         "test/crypto.ts",
         "test/dgram.ts",
         "test/global.ts",
         "test/http.ts",
         "test/http2.ts",
         "test/net.ts",
+        "test/os.ts",
+        "test/path.ts",
         "test/perf_hooks.ts",
+        "test/querystring.ts",
         "test/readline.ts",
+        "test/repl.ts",
         "test/stream.ts",
+        "test/tls.ts",
         "test/tty.ts",
         "test/util.ts",
-        "test/repl.ts",
         "test/worker_threads.ts",
-        "test/zlib.ts",
-        "test/tls.ts"
+        "test/zlib.ts"
     ],
     "compilerOptions": {
         "module": "commonjs",
